+++ R/S3-methods.R
@@ -44,23 +44,41 @@ print.sjmisc.frq <- function(x, ...) {
 #' @importFrom crayon blue
 #' @importFrom cli cat_line
 #' @export
-print.sjmisc.descr <- function(x, ...) {
+print.sjmisc_descr <- function(x, ...) {
+  cat("\n")
+  cli::cat_line(crayon::blue("## Basic descriptive statistics\n"))
+  print_descr_helper(x, ...)
+}
 
+print_descr_helper <- function(x, ...) {
   digits <- 2
 
   # do we have digits argument?
   add.args <- lapply(match.call(expand.dots = F)$`...`, function(x) x)
   if ("digits" %in% names(add.args)) digits <- eval(add.args[["digits"]])
 
-
-  cat("\n")
-  cli::cat_line(crayon::blue("## Basic descriptive statistics\n"))
   # round values
   x[, c(5:10, 14:15)] <- round(x[, c(5:10, 14:15)], digits = digits)
   # print frq-table
   print.data.frame(x, ..., row.names = FALSE)
 }
 
+#' @importFrom crayon blue cyan italic
+#' @importFrom cli cat_line
+#' @export
+print.sjmisc_grpdescr <- function(x, ...) {
+  cat("\n")
+  cli::cat_line(crayon::blue("## Basic descriptive statistics"))
+
+  purrr::walk(x, function(.x) {
+    # print title for grouping
+    cli::cat_line(crayon::cyan(crayon::italic(
+      sprintf("\nGrouped by:\n%s", attr(.x, "group", exact = TRUE))
+    )))
+
+    print_descr_helper(.x, ...)
+  })
+}
 
 #' @importFrom purrr map_df
 #' @importFrom dplyr n_distinct filter